requirements.txt
streamlit
pandas
numpy
requests
altair
google-generativeai>=0.7.0
supabase
